// src/app/api/api.ts
export class BookingDTO implements IBookingDTO {
    id?: string | undefined;
    number?: number | undefined;
    startDate?: Date | undefined;
    endDate?: Date | undefined;
    totalPrice?: number | undefined;
    status?: string | undefined;
    periodCoefficientId?: string | undefined;
    depotId?: string | undefined;
    vehicleID?: string | undefined;
    clientId?: string | undefined;
    vehicle?: VehicleDTO | undefined;
    periodCoefficient?: PeriodCoefficientDTO | undefined;
    ageCoefficient?: AgeCoefficientDTO | undefined;
    client?: ClientDTO | undefined;
    depot?: DepotDTO | undefined;
    optionBookingBookings?: OptionBookingBookingDTO[] | undefined;

    constructor(data?: IBookingDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.number = data["Number"];
            this.startDate = data["StartDate"] ? new Date(data["StartDate"].toString()) : <any>undefined;
            this.endDate = data["EndDate"] ? new Date(data["EndDate"].toString()) : <any>undefined;
            this.totalPrice = data["TotalPrice"];
            this.status = data["Status"];
            this.periodCoefficientId = data["PeriodCoefficientId"];
            this.depotId = data["DepotId"];
            this.vehicleID = data["VehicleID"];
            this.clientId = data["ClientId"];
            this.vehicle = data["Vehicle"] ? VehicleDTO.fromJS(data["Vehicle"]) : <any>undefined;
            this.periodCoefficient = data["PeriodCoefficient"] ? PeriodCoefficientDTO.fromJS(data["PeriodCoefficient"]) : <any>undefined;
            this.ageCoefficient = data["AgeCoefficient"] ? AgeCoefficientDTO.fromJS(data["AgeCoefficient"]) : <any>undefined;
            this.client = data["Client"] ? ClientDTO.fromJS(data["Client"]) : <any>undefined;
            this.depot = data["Depot"] ? DepotDTO.fromJS(data["Depot"]) : <any>undefined;
            if (Array.isArray(data["OptionBookingBookings"])) {
                this.optionBookingBookings = [] as any;
                for (let item of data["OptionBookingBookings"])
                    this.optionBookingBookings!.push(OptionBookingBookingDTO.fromJS(item));
            }
        }
    }

    static fromJS(data: any): BookingDTO {
        data = typeof data === 'object' ? data : {};
        let result = new BookingDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Number"] = this.number;
        data["StartDate"] = this.startDate ? this.startDate.toISOString() : <any>undefined;
        data["EndDate"] = this.endDate ? this.endDate.toISOString() : <any>undefined;
        data["TotalPrice"] = this.totalPrice;
        data["Status"] = this.status;
        data["PeriodCoefficientId"] = this.periodCoefficientId;
        data["DepotId"] = this.depotId;
        data["VehicleID"] = this.vehicleID;
        data["ClientId"] = this.clientId;
        data["Vehicle"] = this.vehicle ? this.vehicle.toJSON() : <any>undefined;
        data["PeriodCoefficient"] = this.periodCoefficient ? this.periodCoefficient.toJSON() : <any>undefined;
        data["AgeCoefficient"] = this.ageCoefficient ? this.ageCoefficient.toJSON() : <any>undefined;
        data["Client"] = this.client ? this.client.toJSON() : <any>undefined;
        data["Depot"] = this.depot ? this.depot.toJSON() : <any>undefined;
        if (Array.isArray(this.optionBookingBookings)) {
            data["OptionBookingBookings"] = [];
            for (let item of this.optionBookingBookings)
                data["OptionBookingBookings"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IBookingDTO {
    id?: string | undefined;
    number?: number | undefined;
    startDate?: Date | undefined;
    endDate?: Date | undefined;
    totalPrice?: number | undefined;
    status?: string | undefined;
    periodCoefficientId?: string | undefined;
    depotId?: string | undefined;
    vehicleID?: string | undefined;
    clientId?: string | undefined;
    vehicle?: VehicleDTO | undefined;
    periodCoefficient?: PeriodCoefficientDTO | undefined;
    ageCoefficient?: AgeCoefficientDTO | undefined;
    client?: ClientDTO | undefined;
    depot?: DepotDTO | undefined;
    optionBookingBookings?: OptionBookingBookingDTO[] | undefined;
}

export class VehicleDTO implements IVehicleDTO {
    id?: string | undefined;
    year?: number | undefined;
    mileage?: number | undefined;
    modelId?: string | undefined;
    plateNumber?: string | undefined;
    chassisNumber?: string | undefined;
    opinionAverage?: number | undefined;
    optionNames?: string[] | undefined;
    optionVehicleVehicles?: OptionVehicleVehicleDTO[] | undefined;
    model?: ModelDTO | undefined;
    maintenances?: MaintenanceDTO[] | undefined;

    constructor(data?: IVehicleDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.year = data["Year"];
            this.mileage = data["Mileage"];
            this.modelId = data["ModelId"];
            this.plateNumber = data["PlateNumber"];
            this.chassisNumber = data["ChassisNumber"];
            this.opinionAverage = data["OpinionAverage"];
            if (Array.isArray(data["OptionNames"])) {
                this.optionNames = [] as any;
                for (let item of data["OptionNames"])
                    this.optionNames!.push(item);
            }
            if (Array.isArray(data["OptionVehicleVehicles"])) {
                this.optionVehicleVehicles = [] as any;
                for (let item of data["OptionVehicleVehicles"])
                    this.optionVehicleVehicles!.push(OptionVehicleVehicleDTO.fromJS(item));
            }
            this.model = data["Model"] ? ModelDTO.fromJS(data["Model"]) : <any>undefined;
            if (Array.isArray(data["Maintenances"])) {
                this.maintenances = [] as any;
                for (let item of data["Maintenances"])
                    this.maintenances!.push(MaintenanceDTO.fromJS(item));
            }
        }
    }

    static fromJS(data: any): VehicleDTO {
        data = typeof data === 'object' ? data : {};
        let result = new VehicleDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Year"] = this.year;
        data["Mileage"] = this.mileage;
        data["ModelId"] = this.modelId;
        data["PlateNumber"] = this.plateNumber;
        data["ChassisNumber"] = this.chassisNumber;
        data["OpinionAverage"] = this.opinionAverage;
        if (Array.isArray(this.optionNames)) {
            data["OptionNames"] = [];
            for (let item of this.optionNames)
                data["OptionNames"].push(item);
        }
        if (Array.isArray(this.optionVehicleVehicles)) {
            data["OptionVehicleVehicles"] = [];
            for (let item of this.optionVehicleVehicles)
                data["OptionVehicleVehicles"].push(item.toJSON());
        }
        data["Model"] = this.model ? this.model.toJSON() : <any>undefined;
        if (Array.isArray(this.maintenances)) {
            data["Maintenances"] = [];
            for (let item of this.maintenances)
                data["Maintenances"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IVehicleDTO {
    id?: string | undefined;
    year?: number | undefined;
    mileage?: number | undefined;
    modelId?: string | undefined;
    plateNumber?: string | undefined;
    chassisNumber?: string | undefined;
    opinionAverage?: number | undefined;
    optionNames?: string[] | undefined;
    optionVehicleVehicles?: OptionVehicleVehicleDTO[] | undefined;
    model?: ModelDTO | undefined;
    maintenances?: MaintenanceDTO[] | undefined;
}

export class PeriodCoefficientDTO implements IPeriodCoefficientDTO {
    id?: string | undefined;
    startDate?: Date | undefined;
    endDate?: Date | undefined;
    percentage?: number | undefined;

    constructor(data?: IPeriodCoefficientDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.startDate = data["StartDate"] ? new Date(data["StartDate"].toString()) : <any>undefined;
            this.endDate = data["EndDate"] ? new Date(data["EndDate"].toString()) : <any>undefined;
            this.percentage = data["Percentage"];
        }
    }

    static fromJS(data: any): PeriodCoefficientDTO {
        data = typeof data === 'object' ? data : {};
        let result = new PeriodCoefficientDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["StartDate"] = this.startDate ? this.startDate.toISOString() : <any>undefined;
        data["EndDate"] = this.endDate ? this.endDate.toISOString() : <any>undefined;
        data["Percentage"] = this.percentage;
        return data; 
    }
}

export interface IPeriodCoefficientDTO {
    id?: string | undefined;
    startDate?: Date | undefined;
    endDate?: Date | undefined;
    percentage?: number | undefined;
}
// ...
export class ClientDTO implements IClientDTO {
    id?: string | undefined;
    surname?: string | undefined;
    firstName?: string | undefined;
    driverLicenseNumber?: string | undefined;
    birthDate?: Date | undefined;
    email?: string | undefined;
    password?: string | undefined;

    constructor(data?: IClientDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.surname = data["Surname"];
            this.firstName = data["FirstName"];
            this.driverLicenseNumber = data["DriverLicenseNumber"];
            this.birthDate = data["BirthDate"] ? new Date(data["BirthDate"].toString()) : <any>undefined;
            this.email = data["Email"];
            this.password = data["Password"];
        }
    }

    static fromJS(data: any): ClientDTO {
        data = typeof data === 'object' ? data : {};
        let result = new ClientDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Surname"] = this.surname;
        data["FirstName"] = this.firstName;
        data["DriverLicenseNumber"] = this.driverLicenseNumber;
        data["BirthDate"] = this.birthDate ? this.birthDate.toISOString() : <any>undefined;
        data["Email"] = this.email;
        data["Password"] = this.password;
        return data; 
    }
}

export interface IClientDTO {
    id?: string | undefined;
    surname?: string | undefined;
    firstName?: string | undefined;
    driverLicenseNumber?: string | undefined;
    birthDate?: Date | undefined;
    email?: string | undefined;
    password?: string | undefined;
}